split siteMap into headerSiteMap and footerSiteMap to have different links
